refactored filepaths
