+++ src/restful/routes/index.js
@@ -24,4 +24,4 @@ router.use("*", (req, res) => {
   });
 });
 
-exports.router = router;
+module.exports = router;
