+++ src/restful/routes/opportunitiesRouters.js
@@ -1,15 +1,11 @@
 const { Router } = require("express");
-const OpportunitiesController = require("../controllers/opportunitiesControllers");
+const OpportunitiesController = require("../controllers/opportunitiesController");
 
 const router = Router();
 
 router.get("/", OpportunitiesController.getAllOpportunities);
-// router.get("/open", OpportunitiesController.getOpenOpportunities);
 router.get("/id/:user_id", OpportunitiesController.getAllOpportunitiesByUserId);
 router.get("/status/:status", OpportunitiesController.getOpportunitiesByStatus);
-// router.get("/in_progress", OpportunitiesController.getInProgressOpportunities);
-// router.get("/completed", OpportunitiesController.getCompletedOpportunities);
-// router.get("/:doc_type/:opportunity_id", OpportunitiesController.getOpportunityById); // Make sure this route is correct
 router.delete(
   "/opportunityid/:opportunity_id",
   OpportunitiesController.deleteOpportunityById,
@@ -18,12 +14,6 @@ router.get(
   "/opportunityid/:opportunity_id",
   OpportunitiesController.getOpportunityById,
 );
-
-// GET endpoint to retrieve individual opportunity by ID
-// router.get("/:doc_type/:opportunity_id", OpportunitiesController.getOpportunityById);
-
-// POST endpoint
-// router.post("/", OpportunitiesController.createOpportunity);
 
 // POST endpoint for creating opportunities
 router.post("/", async (req, res) => {
@@ -63,4 +53,4 @@ router.put(
   OpportunitiesController.updateOpportunityById,
 ); // New route for update
 
-exports.router = router;
+module.exports = router;
